moved about page to views
